+++ app/exceptions/invalid_constant_exception.rb
@@ -1,0 +1,5 @@
+class InvalidConstantException < StandardError
+  def initialize(m)
+    super("The code '#{m}' is not permitted")
+  end
+end
