+++ app/stack/stack_constant.rb
@@ -1,0 +1,6 @@
+class StackConstant
+  attr_reader :id
+  def initialize(id)
+    @id = id
+  end
+end
