+++ app/stack/stack_indirection.rb
@@ -1,0 +1,10 @@
+class StackIndirection
+  attr_reader :id
+  def initialize(id)
+    @id = id
+  end
+
+  def method_missing(name, *args, &block)
+    StackReference.new(@id, name.to_s)
+  end
+end
