+++ app/stack/stack_reference.rb
@@ -1,0 +1,7 @@
+class StackReference
+  attr_reader :id, :member
+  def initialize(id, member)
+    @id = id
+    @member = member
+  end
+end
